requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-ollama
faiss-cpu
python-dotenv